+++ probabilityModel/Model.java
@@ -1,5 +1,5 @@
 package probabilityModel;
 
 public interface Model {
-	int nextDigit();
+	byte nextDigit();
 }
