+++ simulator/ICGenerator.java
@@ -6,39 +6,39 @@ public class ICGenerator{
 	public ICGenerator(Model probModel){
 		_model=probModel;
 	}
-	public int[] getIC1D(int latticeSize){
-		int[] newIC = new int[latticeSize];
+	public byte[] getIC1D(int latticeSize){
+		byte[] newIC = new byte[latticeSize];
 		populate(newIC,latticeSize);
 		return newIC;
 	}
-	public int[][] getIC2D(int latticeSize){
-		int[][] newIC = new int[latticeSize][latticeSize];
+	public byte[][] getIC2D(int latticeSize){
+		byte[][] newIC = new byte[latticeSize][latticeSize];
 		populate(newIC,latticeSize);
 		return newIC;
 	}
-	public int[][][] getIC3D(int latticeSize){
-		int[][][] newIC = new int[latticeSize][latticeSize][latticeSize];
+	public byte[][][] getIC3D(int latticeSize){
+		byte[][][] newIC = new byte[latticeSize][latticeSize][latticeSize];
 		populate(newIC,latticeSize);
 		return newIC;
 	}
-	public int[][][][] getIC4D(int latticeSize){
-		int[][][][] newIC = new int[latticeSize][latticeSize][latticeSize][latticeSize];
+	public byte[][][][] getIC4D(int latticeSize){
+		byte[][][][] newIC = new byte[latticeSize][latticeSize][latticeSize][latticeSize];
 		populate(newIC,latticeSize);
 		return newIC;
 	}
-	private void populate(int[] IC, int latticeSize){
+	private void populate(byte[] IC, int latticeSize){
 		for(int cols=0;cols<latticeSize;cols++){
 			IC[cols]=getValue();
 		}
 	}
-	private void populate(int[][] IC, int latticeSize){
+	private void populate(byte[][] IC, int latticeSize){
 		for(int rows=0;rows<latticeSize;rows++){
 			for(int cols=0;cols<latticeSize;cols++){
 				IC[rows][cols]=getValue();
 			}
 		}
 	}
-	private void populate(int[][][] IC, int latticeSize){
+	private void populate(byte[][][] IC, int latticeSize){
 		for(int planes=0; planes<latticeSize;planes++){
 			for(int rows=0;rows<latticeSize;rows++){
 				for(int cols=0;cols<latticeSize;cols++){
@@ -47,7 +47,7 @@ public class ICGenerator{
 			}
 		}
 	}
-	private void populate(int[][][][] IC, int latticeSize){
+	private void populate(byte[][][][] IC, int latticeSize){
 		for(int cubes=0;cubes<latticeSize;cubes++){
 			for(int planes=0; planes<latticeSize;planes++){
 				for(int rows=0;rows<latticeSize;rows++){
@@ -58,7 +58,7 @@ public class ICGenerator{
 			}
 		}
 	}
-	public int getValue(){
+	public byte getValue(){
 		return _model.nextDigit();
 	}
 }
